deepCrawler/classifier/requirements.txt
kafka-python
transformers
torch
boto3
iocextract
elasticsearch
